+++ packages/core/src/store/mad-config/mad-config.ts
@@ -1,10 +1,13 @@
 import { create } from "zustand";
 import { devtools } from "zustand/middleware";
-import { MadConfig } from "../../types";
+import {Environment, EnvironmentContextualConfig, MadConfig} from "../../types";
+import { createEnvironmentProxy } from "../../utils/createEnvironmentProxy";
+import { useMemo } from "react";
 
 type MadConfigState = {
     config: MadConfig | null;
     setConfig: (config: MadConfig) => void;
+    setEnvironment: (env: Environment) => void;
 };
 
 const useMadConfigStore = create<MadConfigState>()(
@@ -12,6 +15,11 @@ const useMadConfigStore = create<MadConfigState>()(
         set => ({
             config: null,
             setConfig: config => set(() => ({ config: config })),
+            setEnvironment: env => {
+                const newConfig = getPureConfig();
+                newConfig.currentEnvironment = env;
+                set(() => ({ config: newConfig }))
+            },
         }),
         { name: "core/config" },
     ),
@@ -20,13 +28,21 @@ const useMadConfigStore = create<MadConfigState>()(
 const MAD_CONFIG_NOT_FOUND_ERROR =
     "Mad config has not been provided! Make sure you use 'createCoreStackNavigator' to provide your config";
 
-export const useMadConfig = (): MadConfig => {
+export const useMadConfig = (): EnvironmentContextualConfig => {
     const config = useMadConfigStore().config;
     if (!config) throw new Error(MAD_CONFIG_NOT_FOUND_ERROR);
-    return config;
+    return useMemo(() => {
+        return createEnvironmentProxy(config.currentEnvironment);
+    }, [config, config.currentEnvironment]);
 };
 
-export const getConfig = (): MadConfig => {
+export const getConfig = (): EnvironmentContextualConfig => {
+    const config = useMadConfigStore.getState().config;
+    if (!config) throw new Error(MAD_CONFIG_NOT_FOUND_ERROR);
+    return createEnvironmentProxy(config.currentEnvironment);
+};
+
+export const getPureConfig = (): MadConfig => {
     const config = useMadConfigStore.getState().config;
     if (!config) throw new Error(MAD_CONFIG_NOT_FOUND_ERROR);
     return config;
@@ -34,4 +50,8 @@ export const getConfig = (): MadConfig => {
 
 export const setConfig = (config: MadConfig) => {
     useMadConfigStore.getState().setConfig(config);
+};
+
+export const setEnvironment = (env: Environment) => {
+    useMadConfigStore.getState().setEnvironment(env);
 };