+++ packages/core/src/types.ts
@@ -1,4 +1,3 @@
-import { EnvironmentContextProps } from "@equinor/mad-components";
 import { ImageSourcePropType } from "react-native";
 import { Language } from "./store/types";
 import { AppInsightsInitConfig } from "@equinor/mad-insights";
@@ -7,17 +6,18 @@ export type MadConfig = {
     /**
      * Version of the app. Will be displayed in the about screen, and will be used for release notes
      */
-    appVersion: string;
+    appVersion: EnvironmentValues<string>;
     /**
      * service portal name of the app. Will be used to find the correct resource for service messages and release notes.
      * @see https://web-mad-service-portal-web-prod.radix.equinor.com/
      */
-    servicePortalName: string;
+    servicePortalName: EnvironmentValues<string>;
+
     /**
      * Current environment. Will be used for environment banner, as well as getting the correct resource for service messages and release notes
      */
-    environment: EnvironmentContextProps["environment"];
-    language: {
+    currentEnvironment: Environment;
+    language: EnvironmentValues<{
         /**
          * Supported languages of the app.
          */
@@ -30,9 +30,9 @@ export type MadConfig = {
         /**
          * Core navigates to a language selection screen by default if needed. Set this to true if you want to override this behaviour
          */
-        skipOnboarding?: boolean
-    };
-    authentication: {
+        skipOnboarding?: boolean;
+    }>;
+    authentication: EnvironmentValues<{
         /**
          * Client Id of the application. Used for login.
          * You can find your application's client Id in your application's
@@ -59,9 +59,9 @@ export type MadConfig = {
          * available scopes in your application's App registration in Azure.
          * @see https://portal.azure.com/#view/Microsoft_AAD_RegisteredApps/ApplicationsListBlade
          */
-        scopes?: string[];
-    };
-    login: {
+        scopes: string[];
+    }>;
+    login: EnvironmentValues<{
         /**
          * Title of the app. Used in login screen
          */
@@ -70,12 +70,12 @@ export type MadConfig = {
          * App logo. Used in login screen
          */
         logo: ImageSourcePropType;
-    };
+    }>;
     /**
      * App insights config used for initializing application insights service(s)
      */
-    applicationInsights: AppInsightsInitConfig;
-    about?: {
+    applicationInsights: EnvironmentValues<AppInsightsInitConfig>;
+    about?: EnvironmentValues<{
         /**
          * Endpoints used by the app
          */
@@ -84,11 +84,13 @@ export type MadConfig = {
          * Build number of the app.
          */
         buildNumber: string;
-    };
-    serviceNow?: {
-        //TODO
-        whatever: string;
-    };
+    }>;
+    /**
+     * ServiceNow Configuration Item (cmdb_ci) of the app.
+     * Will be used to create ServiceNow tickets for the correlating application.
+     * @see https://portal.azure.com/#@StatoilSRM.onmicrosoft.com/resource/subscriptions/2ad1b087-ffb8-4cf3-bcc2-8caeebfcd3f3/resourceGroups/mad-test/providers/Microsoft.Storage/storageAccounts/madtest62alhixjcodkm/storagebrowser
+     */
+    serviceNow?: EnvironmentValues<string>;
 };
 
 export type CoreStackParamListBase = {
@@ -105,3 +107,11 @@ export type CoreStackParamListBase = {
 };
 
 export type Environment = "dev" | "test" | "qa" | "prod";
+
+export type EnvironmentValues<T> = Partial<Record<Environment, T>> | T;
+
+export type WithoutEnvironmentOptionValues<TToken> = {
+    [K in keyof TToken]: TToken[K] extends EnvironmentValues<infer U> ? U : TToken[K];
+};
+
+export type EnvironmentContextualConfig = WithoutEnvironmentOptionValues<MadConfig>;